sourcery automatically code beautify

+++ stdplugins/log_pms.py
@@ -22,8 +22,8 @@ async def monito_p_m_s(event):
         chat = await event.get_chat()
         if chat.id not in NO_PM_LOG_USERS and chat.id != borg.uid:
             try:
-                e = await borg.get_entity(int(Config.PM_LOGGR_BOT_API_ID))
-                fwd_message = await borg.forward_messages(
+                e = await event.client.get_entity(int(Config.PM_LOGGR_BOT_API_ID))
+                fwd_message = await event.client.forward_messages(
                     e,
                     event.message,
                     silent=True
@@ -42,10 +42,12 @@ async def approve_p_m(event):
         return
     reason = event.pattern_match.group(1)
     chat = await event.get_chat()
-    if Config.NC_LOG_P_M_S:
-        if event.is_private:
-            if chat.id not in NO_PM_LOG_USERS:
-                NO_PM_LOG_USERS.append(chat.id)
-                await event.edit("Won't Log Messages from this chat")
-                await asyncio.sleep(3)
-                await event.delete()
+    if (
+        Config.NC_LOG_P_M_S
+        and event.is_private
+        and chat.id not in NO_PM_LOG_USERS
+    ):
+        NO_PM_LOG_USERS.append(chat.id)
+        await event.edit("Won't Log Messages from this chat")
+        await asyncio.sleep(3)
+        await event.delete()
